Separate unit and functional tests in cml.
