+++ bin/cml/comment/create.e2e.test.js
@@ -1,0 +1,41 @@
+const { exec } = require('../../../src/utils');
+const fs = require('fs').promises;
+
+describe('Comment integration tests', () => {
+  const path = 'comment.md';
+
+  afterEach(async () => {
+    try {
+      await fs.unlink(path);
+    } catch (err) {}
+  });
+
+  test('cml send-comment to specific repo', async () => {
+    const {
+      TEST_GITHUB_REPO: repo,
+      TEST_GITHUB_TOKEN: token,
+      TEST_GITHUB_SHA: sha
+    } = process.env;
+
+    const report = `## Test Comment Report specific`;
+
+    await fs.writeFile(path, report);
+    await exec(
+      `node ./bin/cml.js send-comment --repo=${repo} --token=${token} --commit-sha=${sha} ${path}`
+    );
+  });
+
+  test('cml send-comment to current repo', async () => {
+    const report = `## Test Comment`;
+
+    await fs.writeFile(path, report);
+    await exec(`node ./bin/cml.js send-comment ${path}`);
+  });
+
+  test('cml send-comment --publish to current repo', async () => {
+    const report = `## Test Comment\n![](assets/logo.png)`;
+
+    await fs.writeFile(path, report);
+    await exec(`node ./bin/cml.js send-comment --publish ${path}`);
+  });
+});
