+++ bin/cml/comment/create.test.js
@@ -1,15 +1,6 @@
 const { exec } = require('../../../src/utils');
-const fs = require('fs').promises;
 
 describe('Comment integration tests', () => {
-  const path = 'comment.md';
-
-  afterEach(async () => {
-    try {
-      await fs.unlink(path);
-    } catch (err) {}
-  });
-
   test('cml send-comment --help', async () => {
     const output = await exec(`node ./bin/cml.js send-comment --help`);
 
@@ -43,34 +34,5 @@ describe('Comment integration tests', () => {
         --rm-watermark            Avoid watermark; CML needs a watermark to be able to
                                   distinguish CML comments from others       [boolean]"
     `);
-  });
-
-  test('cml send-comment to specific repo', async () => {
-    const {
-      TEST_GITHUB_REPO: repo,
-      TEST_GITHUB_TOKEN: token,
-      TEST_GITHUB_SHA: sha
-    } = process.env;
-
-    const report = `## Test Comment Report specific`;
-
-    await fs.writeFile(path, report);
-    await exec(
-      `node ./bin/cml.js send-comment --repo=${repo} --token=${token} --commit-sha=${sha} ${path}`
-    );
-  });
-
-  test('cml send-comment to current repo', async () => {
-    const report = `## Test Comment`;
-
-    await fs.writeFile(path, report);
-    await exec(`node ./bin/cml.js send-comment ${path}`);
-  });
-
-  test('cml send-comment --publish to current repo', async () => {
-    const report = `## Test Comment\n![](assets/logo.png)`;
-
-    await fs.writeFile(path, report);
-    await exec(`node ./bin/cml.js send-comment --publish ${path}`);
   });
 });
