+++ jest.config.e2e.js
@@ -1,0 +1,5 @@
+// jest config for e2e tests.
+module.exports = {
+  testRegex: 'e2e.test.js$',
+  testTimeout: 600000
+};
