+++ src/analytics.e2e.test.js
